Issue #1404: Refactoring in creatorUtils classes
